Fixed bug where active skater number would revert to default on page reload
When reloading the penalty timer page while a skater number was active, the menu selection would revert to the default of no skater being selected. When reloading the page it was also possible to sometimes get into a state where the selection would flicker as the menu would try to update between two different values.
Now when the menu is initialized, check the server's idea of who the current skater is, update the selection from that, and don't incorrectly set the skater number to an empty string.