Added skater's lineup role as a letter after their number in the selection list
To make it easier to spot blockers for the current lineup in the skater number selection list, their fielding now appears after the number as B, P, or J. The sort order of the list remains the same. It would be easy to move them all to the top or bottom, but that seems possibly confusing because on a phone the list may require scrolling, so you might not know whether to scroll up or down. Trying it out this way for now.