Fixes to make old pbt interface work in latest scoreboard release
I've merged down the latest CRG scoreboard 2025.3 changes which include a new penalty box timer interface. It uses a different UI and concept than the one I had added, so for people who prefer the original one, I've made some changes to allow them to co-exist.

Server side changes:
Merging the java server side changes went pretty smoothly. At this point they're co-existing without getting in each others way. I haven't made any attempt to change the original pbt to use the new clocks and BoxTrip class methods that the new one uses. This means that using both at the same time is probably not a great idea, since the clocks from the new pbt won't be reflected in the original pbt and vice versa. Think of it as different sets of stop watches.

UI side changes:
The link for the original pbt appears at the bottom of the list on the main menu page so that it's easy to spot.
I've moved the html and javascript part of the code to live under nso/pbt2, because the new one from the main branch now has the name nso/pbt.
The original pbt depends on jquery-mobile, which has been removed in the latest CRG release, so I've added that back into my branch.
The WS.js module, with websocket related functions, changed substantially in the new release in ways that break the original pbt. For the time being, instead of porting to use the new module I've restored the old WS.js with new name WS_old.js, and am continuing to use it in the original pbt.