Added a rule to cause timeout to automatically stop when starting a jam
Added a rule under Timeout section called StopTOOnJamStart, disabled by default to match behavior of official release, that causes a running timeout to automatically be stopped when a jam is started. This prevents accidentally having a timeout running at the same time as a jam if the user did not remember to hit the button to end timeout.